moved util func where it can be accessed by all the primitives
